python-functions/requirements.txt
fastapi
uvicorn
rembg
pillow
requests
supabase
onnxruntime-gpu
python-multipart